Allow creating app in existing empty dir
I tried `yarn create redwood-app .` in an empty directory where I wanted the files generated, and got an error saying the dir already exists.  
I think it should be OK to use an existing directory if it's empty 🙂 

I have also moved the error out of the Listr task itself, because it seems Listr is not actually stopping execution of the following tasks. At least, it continues outputting the next tasks' titles, which is confusing:

I'd say an early exit for this particular error is acceptable?